Create two different repositories, one for dynamodb and one for in-memory store.
